+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Represents the **NuGet** versions.
 
+## v1.0.7
+- *Fixed:* Event Outbox dequeue and send publishing failure will log and continue (allowing retry); versus fail process.
+- *Enhancement:* Template updated to show code (commented out) for Azure Service Bus event publishing.
+
 ## v1.0.6
 - *Enhancement:* Updated `DbEx` dependency to latest.